api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic==2.10.4
pydantic-settings==2.6.1
PyJWT==2.10.1
slowapi==0.1.9
stripe==11.6.0
supabase==2.8.1
